+++ qc-application/src/main/java/com/eager/questioncloud/application/api/subscribe/implement/SubscribeProcessor.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.application.api.subscribe;
+package com.eager.questioncloud.application.api.subscribe.implement;
 
 import com.eager.questioncloud.core.domain.creator.infrastructure.CreatorRepository;
 import com.eager.questioncloud.core.domain.subscribe.infrastructure.SubscribeRepository;
